requirements.txt
-r base.txt

black==23.3.0
django-extensions==3.2.1
django-silk==5.0.3
ipython==8.12.0
isort==5.12
mkdocs-gen-files==0.5.0
mkdocs-literate-nav==0.6.0
mkdocs-material==9.1.8
mkdocs-section-index==0.3.5
mkdocstrings[python]==0.21.2
pydot==1.4.2
pyparsing==3.0.9
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-xdist[psutil]==3.2.1
ruff
werkzeug==2.3.7
